+++ src/app/controllers/processo_seletivo.js
@@ -1,0 +1,59 @@
+const ProcessoSeletivoSchama = require('./../models/processo_seletivo')
+
+class ProcessoSeletivo {
+
+    getWithParams(req, res) {
+
+        let limit = parseInt(req.query.limit)
+        let query = {}
+        let page = req.query.page
+        let skip = limit * (page - 1)
+        let { columnSort, valueSort } = req.query
+
+        ProcessoSeletivoSchama
+            .find(query)
+            .sort([[columnSort, valueSort]])
+            .skip(skip)
+            .limit(limit)
+            .exec((err, data) => {
+                if (err) {
+                    res.status(500).json({ message: 'Houve um erro ao processar sua requisição', err: err })
+                } else {
+                    ProcessoSeletivoSchama
+                        .find(query)
+                        .exec((err, count) => {
+                            let totalDocuments = count.length
+                            if (err) {
+                                res.status(500).json({ message: 'Houve um erro ao processar sua requisição', err: err })
+                            } else {
+                                if (totalDocuments > 0) {
+                                    res.status(200).json({
+                                        message: 'Dados recuperados com sucesso',
+                                        data: data,
+                                        page: page,
+                                        limit: limit,
+                                        count: totalDocuments,
+                                    })
+                                } else {
+                                    res.status(204).json({
+                                        message: 'Não há dados para serem exibidos'
+                                    })
+                                }
+                            }
+                        })
+                }
+            })
+    }
+
+    create(req, res) {
+        ProcessoSeletivoSchama.create(req.body, (err, data) => {
+            if (err) {
+                res.status(500).json({ message: 'Houve um erro ao processar sua requisição', error: err })
+            } else {
+                res.status(201).json({ message: 'Processo Seletivo criado com sucesso', data: data })
+            }
+        })
+    }
+}
+
+module.exports = new ProcessoSeletivo()