+++ src/app/routes/private/routes/publicacoes.private.js
@@ -1,0 +1,8 @@
+const express = require('express')
+const route = express.Router()
+const Publicacoes = require('./../../../controllers/publicacoes')
+
+route.post('/criar', Publicacoes.create)
+route.get('/listar-todos', Publicacoes.getWithParams)
+
+module.exports = route