+++ CHANGELOG.md
@@ -4,3 +4,9 @@ Log de alterações da aplicação
 
 ### Versão 1.0
 * **Contatos:** Implementado CRUD completo e separado em rotas públicas e privadas.
+* **Processo Seletivo:** Criado modelo, controller com os métodos de listar todos e criar, validador de título único, criação do arquivo de rotas para o módulo e alteração no arquivo de rotas privadas para inserir a rota de Processo Seletivo.
+* **Integrantes:** Implementando create e getall nas rotas privadas.
+* **Publicações:** Implementado o método Create e getWithParams para listar todos, validador de título único.
+* **Categorias:** Reaproveitando a implementação de categories da AsiloWebAPI para implementar os métodos de criar e listar as categorias, validador de nome único.
+* **Eventos:** Desenvolvido o método Create e GetWithParams.
+* **Sobre:** Desenvolvido o metodo create no módulo sobre.